+++ src/components/modal.jsx
@@ -1,0 +1,28 @@
+import React from 'react'
+import { useState } from 'react';
+import { Modal, Button } from "react-bootstrap";
+
+
+
+
+const ModalR = (props) => {
+    
+  return (
+    <Modal key={props.local_direccion} show={props.show} onHide={props.handleClose}>
+        <Modal.Header closeButton>
+          <Modal.Title>Modal heading</Modal.Title>
+        </Modal.Header>
+        <Modal.Body>{props.local_direccion}</Modal.Body>
+        <Modal.Footer>
+          <Button variant="secondary" onClick={props.handleClose}>
+            Close
+          </Button>
+          <Button variant="primary" onClick={props.handleClose}>
+            Save Changes
+          </Button>
+        </Modal.Footer>
+      </Modal>
+  )
+}
+
+export default ModalR
